+++ xuecheng-edu-content/xuecheng-edu-content-model/src/main/java/com/xuecheng/content/model/dto/CourseBaseInfoDto.java
@@ -26,12 +26,12 @@ public class CourseBaseInfoDto extends CourseBase {
     /**
      * 价格
      */
-    private Float price;
+    private Double price;
 
     /**
      * 原价
      */
-    private Float originalPrice;
+    private Double originalPrice;
 
     /**
      * 咨询qq
